[era] Factor out MetadataGenerator
Also validate parameters before writing out the metadata
