+++ src/era/mod.rs
@@ -7,3 +7,6 @@ pub mod restore;
 pub mod superblock;
 pub mod writeset;
 pub mod xml;
+
+#[cfg(feature = "devtools")]
+pub mod metadata_generator;
